# openapi.yaml
openapi: 3.0.3
info:
  title: ShadowCheck API
  description: |
    ShadowCheck is a wireless network threat detection platform that analyzes WiFi, Bluetooth,
    and cellular observations to detect potential surveillance devices.

    ## Authentication

    Most endpoints require API key authentication via the `X-API-Key` header or `api_key` query parameter.

    ## Network Types

    - `W`: WiFi
    - `E`: BLE (Bluetooth Low Energy)
    - `B`: Bluetooth
    - `L`: LTE
    - `N`: 5G NR
    - `G`: GSM

    ## Threat Detection

    Networks are scored based on movement patterns, distance ranges, and observation frequency.
    Scores ≥40 are flagged as potential threats.
  version: 1.0.0
  contact:
    name: ShadowCheck
  license:
    name: MIT
servers:
  - url: http://localhost:3001
    description: Local development server
  - url: http://localhost:3001/api
    description: API routes
...
paths:
...
  /api/networks:
    get:
      tags:
        - Networks
      summary: List all networks
      description: Retrieve paginated list of networks with optional filtering and sorting
      operationId: listNetworks
      parameters:
        - name: page
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
          description: Page number
        - name: limit
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 5000
          description: Results per page
        - name: search
          in: query
          schema:
            type: string
          description: Search by SSID or BSSID
        - name: type
          in: query
          schema:
            type: string
            enum: [W, E, B, L, N, G]
          description: Filter by network type
        - name: security
          in: query
          schema:
            type: string
          description: Filter by security/encryption type
        - name: minSignal
          in: query
          schema:
            type: integer
          description: Minimum signal strength (dBm)
        - name: maxSignal
          in: query
          schema:
            type: integer
          description: Maximum signal strength (dBm)
        - name: sort
          in: query
          schema:
            type: string
            enum: [type, ssid, bssid, signal, security, frequency, channel, observations, latitude, longitude, distanceFromHome, accuracy, lastSeen]
            default: lastSeen
          description: Sort column
        - name: order
          in: query
          schema:
            type: string
            enum: [ASC, DESC]
            default: DESC
          description: Sort order
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NetworksListResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/networks/search/{ssid}:
    get:
      tags:
        - Networks
      summary: Search networks by SSID
      description: Search for networks matching the provided SSID
      operationId: searchNetworksBySSID
      parameters:
        - name: ssid
          in: path
          required: true
          schema:
            type: string
          description: SSID to search for
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NetworkSearchResponse'
        '400':
          description: Invalid SSID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/networks/observations/{bssid}:
    get:
      tags:
        - Networks
      summary: Get observations for a network
      description: Retrieve all observations for a specific BSSID
      operationId: getNetworkObservations
      parameters:
        - name: bssid
          in: path
          required: true
          schema:
            type: string
            pattern: '^([0-9A-F]{2}[:-]){5}([0-9A-F]{2})$'
          description: Network BSSID (MAC address)
      responses:
        '200':
          description: Observations retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObservationsResponse'
  /api/networks/tagged:
    get:
      tags:
        - Networks
      summary: Get tagged networks
      description: Retrieve networks with specific user tags
      operationId: getTaggedNetworks
      parameters:
        - name: tag_type
          in: query
          required: true
          schema:
            type: string
            enum: [LEGIT, FALSE_POSITIVE, INVESTIGATE, THREAT]
          description: Tag type to filter by
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 50
          description: Results per page
      responses:
        '200':
          description: Tagged networks retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaggedNetworksResponse'
  /api/tag-network:
    post:
      tags:
        - Networks
      summary: Tag a network
      description: Add or update a tag on a network
      operationId: tagNetwork
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagNetworkRequest'
      responses:
        '200':
          description: Network tagged successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagNetworkResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tag-network/{bssid}:
    delete:
      tags:
        - Networks
      summary: Remove tag from network
      description: Delete a tag from a specific network
      operationId: deleteNetworkTag
      parameters:
        - name: bssid
          in: path
          required: true
          schema:
            type: string
            pattern: '^([0-9A-F]{2}[:-]){5}([0-9A-F]{2})$'
          description: Network BSSID
      responses:
        '200':
          description: Tag removed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  message:
                    type: string
                  bssid:
                    type: string
        '404':
          description: Tag not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/manufacturer/{bssid}:
    get:
      tags:
        - Networks
      summary: Get manufacturer from MAC address
      description: Look up device manufacturer from BSSID/MAC address
      operationId: getManufacturer
      parameters:
        - name: bssid
          in: path
          required: true
          schema:
            type: string
            pattern: '^([0-9A-F]{2}[:-]){5}([0-9A-F]{2})$'
          description: BSSID/MAC address
      responses:
        '200':
          description: Manufacturer information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManufacturerResponse'
  /api/networks/tag-threats:
    post:
      tags:
        - Networks
      summary: Tag multiple networks as threats
      description: Bulk tag multiple networks as threats
      operationId: tagMultipleThreats
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - bssids
              properties:
                bssids:
                  type: array
                  items:
                    type: string
                  description: Array of BSSIDs to tag
                reason:
                  type: string
                  description: Reason for tagging as threats
      responses:
        '200':
          description: Networks tagged
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  message:
                    type: string
                  successCount:
                    type: integer
                  errorCount:
                    type: integer
                  results:
                    type: array
                    items:
                      type: object
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
  schemas:
...
    Network:
      type: object
      properties:
        id:
          type: string
        ssid:
          type: string
        bssid:
          type: string
        type:
          type: string
          enum: [W, E, B, L, N, G]
        security:
          type: string
        encryption:
          type: string
        frequency:
          type: number
          nullable: true
        channel:
          type: integer
          nullable: true
        signal:
          type: integer
          nullable: true
        accuracy:
          type: number
        observations:
          type: integer
        manufacturer:
          type: string
        lastSeen:
          type: integer
          format: int64
          description: Unix timestamp in milliseconds
        timestamp:
          type: integer
          format: int64
        status:
          type: string
          enum: [safe, warning, threat]
        distanceFromHome:
          type: number
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        location:
          type: object
          properties:
            lat:
              type: number
              nullable: true
            lng:
              type: number
              nullable: true
    NetworksListResponse:
      type: object
      properties:
        networks:
          type: array
          items:
            $ref: '#/components/schemas/Network'
        total:
          type: integer
        totalCount:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
    NetworkSearchResponse:
      type: object
      properties:
        ok:
          type: boolean
        query:
          type: string
        count:
          type: integer
        networks:
          type: array
          items:
            type: object
    ObservationsResponse:
      type: object
      properties:
        ok:
          type: boolean
        bssid:
          type: string
        observations:
          type: array
          items:
            $ref: '#/components/schemas/Observation'
        home:
          type: object
          nullable: true
          properties:
            lon:
              type: number
            lat:
              type: number
        count:
          type: integer
    Observation:
      type: object
      properties:
        id:
          type: string
        bssid:
          type: string
        ssid:
          type: string
        type:
          type: string
        encryption:
          type: string
        lat:
          type: number
        lon:
          type: number
        signal:
          type: integer
        time:
          type: integer
          format: int64
        acc:
          type: number
        alt:
          type: number
          nullable: true
        distance_from_home_km:
          type: number
          nullable: true
    TaggedNetworksResponse:
      type: object
      properties:
        ok:
          type: boolean
        tag_type:
          type: string
        networks:
          type: array
          items:
            type: object
            properties:
              bssid:
                type: string
              ssid:
                type: string
              tag_type:
                type: string
              confidence:
                type: number
              notes:
                type: string
              tagged_at:
                type: string
                format: date-time
              updated_at:
                type: string
                format: date-time
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
    TagNetworkRequest:
      type: object
      required:
        - bssid
        - tag_type
        - confidence
      properties:
        bssid:
          type: string
          pattern: '^([0-9A-F]{2}[:-]){5}([0-9A-F]{2})$'
        tag_type:
          type: string
          enum: [LEGIT, FALSE_POSITIVE, INVESTIGATE, THREAT]
        confidence:
          type: number
          minimum: 0
          maximum: 100
        notes:
          type: string
    TagNetworkResponse:
      type: object
      properties:
        ok:
          type: boolean
        tag:
          type: object
    ManufacturerResponse:
      type: object
      properties:
        ok:
          type: boolean
        bssid:
          type: string
        manufacturer:
          type: string
        address:
          type: string
        prefix:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string